Some fixes after compilation with -Wall -Wextra, fixed

+++ examples/ConvertExample/src/WriterRootTreeOPAL.cc
@@ -14,8 +14,8 @@ void WriterRootTreeOPAL::init_branches()
 void WriterRootTreeOPAL::write_event(const GenEvent &evt)
 {
     m_Ievnt=evt.event_number();
-    std::vector<int> beams;
-    for (int i=0;i<evt.particles().size();i++)
+    std::vector<size_t> beams;
+    for (size_t i=0;i<evt.particles().size();i++)
     if (evt.particles().at(i)->status()==4&&std::abs(evt.particles().at(i)->pid())==11)  
     beams.push_back(i);
